# server/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: White-Label School Management System
  version: "3.4.0"
  description: >
    Multi-tenant school management API. Supports SuperAdmin (platform-level)
    and tenant-scoped Admin/Teacher/Student roles.
    v3.4.0: Student role added; attendance correction (PUT /attendance/{recordId});
    link-student-account (PUT /students/{studentId}/link-account);
    reactivate tenant (PUT /super-admin/tenants/{tenantId}/reactivate);
    createTenant now requires an admin block.
servers:
  - url: http://localhost:3000/api
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp]
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Invalid credentials
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
...
    User:
      type: object
      required: [id, name, email, roles]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
            enum: [Teacher, Admin, Student]
          minItems: 1
...
    BulkDeleteRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
          uniqueItems: true
    BulkDeleteResponse:
      type: object
      required: [deleted, failed]
      properties:
        deleted:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            type: object
            required: [id, reason, message]
            properties:
              id:
                type: string
              reason:
                type: string
                enum: [NOT_FOUND, HAS_REFERENCES]
              message:
                type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Missing or invalid token
              timestamp: "2026-02-26T01:00:00Z"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: Insufficient permissions
              timestamp: "2026-02-26T01:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /users:
    get:
      summary: List users
      operationId: listUsers
      tags: [Users]
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [Teacher, Admin, TeacherAdmin]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              examples:
                success:
                  value:
                    users:
                      - id: U123
                        name: John Doe
                        email: john@school1.com
                        roles: [Teacher]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      summary: Create user
      operationId: createUser
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, roles]
              properties:
                name:
                  type: string
                  maxLength: 255
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                roles:
                  type: array
                  items:
                    type: string
                    enum: [Teacher, Admin]
                  minItems: 1
            examples:
              success:
                value:
                  name: Jane Smith
                  email: jane@school1.com
                  password: securepass1
                  roles: [Teacher]
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /users/bulk:
    delete:
      summary: Bulk soft-delete users
      operationId: bulkDeleteUsers
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRequest"
            examples:
              success:
                value:
                  ids: [U001, U002, U003]
      responses:
        "200":
          description: Bulk delete result (partial success allowed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResponse"
              examples:
                partial:
                  value:
                    deleted: [U001, U002]
                    failed:
                      - id: U003
                        reason: HAS_REFERENCES
                        message: "Cannot delete: user has active timeslot assignments or attendance records"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/{id}:
    delete:
      summary: Soft-delete a user
      operationId: deleteUser
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /users/{id}/roles:
    put:
      summary: Update user roles
      operationId: updateUserRoles
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              properties:
                roles:
                  type: array
                  items:
                    type: string
                    enum: [Teacher, Admin, Student]
                  minItems: 1
                  uniqueItems: true
            examples:
              add_admin:
                value:
                  roles: [Teacher, Admin]
      responses:
        "200":
          description: Roles updated
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
              examples:
                success:
                  value:
                    user:
                      id: U124
                      name: John Doe
                      email: john@school1.com
                      roles: [Teacher, Admin]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cannot remove own Admin role — last admin in tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: LASTADMIN
                  message: "Cannot remove Admin role: you are the last admin of this tenant"
                  timestamp: "2026-02-26T07:00:00Z"

  # ─────────────────────────────────────────
  # STUDENTS
  # ─────────────────────────────────────────
...